:sparkles: Rename server to cli and prepare server
