+++ .env
@@ -1,0 +1,1 @@
+MONGO_URL=mongodb+srv://johanna-b:pg.bew3P!CS7R-.@clusterjb.oynux.mongodb.net/drivethru?retryWrites=true&w=majority
