+++ package.json
@@ -3,6 +3,7 @@
   "version": "0.0.0",
   "scripts": {
     "client:dev": "vite",
+    "cli:dev": "ts-node --project tsconfig.server.json src/cli",
     "server:dev": "ts-node --project tsconfig.server.json src/server",
     "server:build": "tsc --project tsconfig.server.json",
     "client:build": "vite build",
