+++ src/utils/messages.ts
@@ -1,4 +1,4 @@
-// export const printPassword = (service: string): void => {
-//   const password = service + "123";
-//   console.log(`The password for ${service} is ${password}`);
-// };
+export const printPassword = (service: string): void => {
+  const password = service + "123";
+  console.log(`The password for ${service} is ${password}`);
+};
